requirements.txt
absl-py==0.11.0
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
boto3==1.16.50
botocore==1.19.50
brewer2mpl==1.4.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
emoji==0.6.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
gensim==3.8.3
ggplot==0.11.5
google-auth==1.24.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pyspellchecker==0.5.5
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smart-open==4.0.1
statsmodels==0.12.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
tornado==6.1
tqdm==4.55.0
traitlets==5.0.5
transformers==4.1.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.3.1
